tests/requirements.txt
assemblyline
assemblyline-service-utilities
pytest
beautifulsoup4
pyyaml
lxml
